+++ log/sequencer.go
@@ -150,7 +150,7 @@ func (s Sequencer) initMerkleTreeFromStorage(ctx context.Context, currentRoot tr
 
 func (s Sequencer) createRootSignature(ctx context.Context, root trillian.SignedLogRoot) (*sigpb.DigitallySigned, error) {
 	trillianSigner := crypto.NewSigner(s.keyManager.SignatureAlgorithm(), s.keyManager)
-	signature, err := trillianSigner.SignLogRoot(root)
+	signature, err := trillianSigner.Sign(crypto.HashLogRoot(root))
 	if err != nil {
 		glog.Warningf("%s: signer failed to sign root: %v", util.LogIDPrefix(ctx), err)
 		return nil, err
